refactor(infinite): update multimedia includes to use common interfaces
