+++ client/gsm/states/scenes/Infinite/InfiniteState.cpp
@@ -15,9 +15,9 @@
 #include "../../../../../common/ECS/entity/Entity.hpp"
 #include "../../../../../common/gsm/IGameStateMachine.hpp"
 #include "../Settings/SettingsState.hpp"
-#include "../../../../../libs/Multimedia/IWindow.hpp"
-#include "../../../../../libs/Multimedia/IEvent.hpp"
-#include "../../../../../libs/Multimedia/IAudio.hpp"
+#include "../../../../../common/interfaces/IWindow.hpp"
+#include "../../../../../common/interfaces/IEvent.hpp"
+#include "../../../../../common/interfaces/IAudio.hpp"
 #include "../../../../../common/InputMapping/IInputProvider.hpp"
 #include "../../../../components/rendering/HitboxRenderComponent.hpp"
 #include "../../../../systems/rendering/AnimationRenderingSystem.hpp"
